+++ src/Magus.DataBuilder/PatchNoteUpdater.cs
@@ -220,8 +220,8 @@ public sealed class PatchNoteUpdater
         return patchNote;
     }
 
-    private static ulong GetPatchTimestamp(KVObject patch)
-        => (ulong)DateTimeOffset.Parse(patch.Children.First(x => x.Name == "patch_date").Value.ToString()!).ToUnixTimeSeconds();
+    private static long GetPatchTimestamp(KVObject patch)
+        => DateTimeOffset.Parse(patch.Children.First(x => x.Name == "patch_date").Value.ToString()!).ToUnixTimeSeconds();
 
     private async Task StorePatchNoteEmbeds()
     {
